Concluding and testing the authors business rules

+++ src/main/java/io/github/kleitonmarques/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.kleitonmarques.libraryapi.service;
 
+import io.github.kleitonmarques.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.kleitonmarques.libraryapi.model.Autor;
 import io.github.kleitonmarques.libraryapi.repository.AutorRepository;
+import io.github.kleitonmarques.libraryapi.repository.LivroRepository;
 import io.github.kleitonmarques.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +16,14 @@ public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator) {
+    public AutorService(AutorRepository repository,
+                        AutorValidator validator,
+                        LivroRepository livroRepository) {
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -39,6 +45,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor que possui livros cadastrados!");
+        }
+
         repository.delete(autor);
     }
 
@@ -56,5 +66,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro (Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
